Build "tar.gz" and "zip" distribution files.
The distribution files are built on the "verify" lifecycle, which is
just before "install", and the zip and tar.gz files are attached to the
installed jar file (ie. they can be found with the jar file both in the
local maven cache, and when installed on a repository like maven local).